refactor(flashcard-session): use dialog-card's native header slots
Session ownership (FSRS queue, card edit/actions, review flush, member
prefs) now lives in one composable, session-controller.ts, provided
once by index.vue and injected by the studying view. That lets
index.vue render the header via dialog-card's header-start/header-end
slots directly instead of teleporting session-header.vue's markup into
a placeholder div, and drop the closed/finished emit round-trip in
favor of calling requestClose/onClosed straight from the controller.

The close/stop button and edit menu are split into their own
components (session-header-close-button.vue, session-header-menu.vue)
since they're now composed directly in index.vue rather than deep in
the studying subtree.
